+++ ramen-kiosk-2022/kiosk_handler.cpp
@@ -27,9 +27,6 @@ bool OrderCategory::Run(void) {
 	cin >> menu;
 	system("cls");
 	switch (menu) {
-	// case 0:
-		// bPrint = true;
-		// break;
 	case 1:
 		AddOrder();
 		break;
