--- ramen-kiosk-2022/menu_item.h
@@ -1,23 +1,0 @@
-#pragma once
-
-#include <string>
-class Object
-{
-private:
-	std::string name;
-	const unsigned int price;
-	unsigned int count;
-
-public:
-	Object(const std::string& name, unsigned int price, unsigned int count);
-	Object(const Object& copy); //재참조
-
-	unsigned int EditCount(int count);
-	unsigned int Calculate(void) const; //전체 가격 계산
-
-	//외부 참조
-	const std::string& Name = name;
-	const unsigned int& Price = price;
-	const unsigned int& Count = count;
-
-};
