+++ ramen-kiosk-2022/order_list.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <string>
+#include "order.h"
+
+class OrderList
+{
+private:
+	Order** rmList;
+	unsigned int count;
+
+public:
+	OrderList(void);
+	~OrderList(void);
+
+	Order* AddOrder(void);
+	void DeleteOrder(unsigned int index);
+
+	Order* GetOrder(unsigned int index);
+
+	const unsigned int& Count = count;
+};
